change password and apply concept creator on each video file.

+++ controllers/videoController.js
@@ -38,9 +38,11 @@ export const postUpload = async(req, res)=>{
     const newVideo = await Video.create({
         fileUrl:path,
         title,
-        description
-    })
-    console.log(newVideo);
+        description,
+        creator:req.user.id
+    });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id))
 }    
 export const videoDetail = async(req, res)=>{
@@ -48,7 +50,7 @@ export const videoDetail = async(req, res)=>{
         params:{id}
     }=req;
     try{
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate('creator');
         res.render('videoDetail',{pageTitle:video.title, video});
     }catch(error){
         console.log(error);
@@ -62,7 +64,12 @@ export const getEditVideo = async(req, res)=>{
     } = req;
     try{
         const video = await Video.findById(id);
-        res.render('editVideo',{pageTitle:`Edit ${video.title}`, video});
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+
+        }else{
+            res.render('editVideo',{pageTitle:`Edit ${video.title}`, video});
+        }
     }catch(error){
         res.redirect(routes.home);
     }
@@ -86,7 +93,13 @@ export const deleteVideo = async(req, res)=>{
         params:{id}
     }=req;
     try{
-        await Video.findOneAndDelete({_id:id});
+        const video = await Video.findById(id);
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+
+        }else{
+            await Video.findOneAndDelete({_id:id});
+        }
     }catch(error){
         console.log(error);
     }
